21.08.14：倒库流程修复

+++ simserver/task/SimCarrierTask.cs
@@ -612,7 +612,7 @@ namespace simtask
                         switch (SORT_STEP)
                         {
                             case SimCarrierSortStepE.获取取货库存位置:
-                                Stock behindstock = PubMaster.Goods.GetStockBehindStockPoint(EndTrack.id, (ushort)(DevStatus.CurrentPoint-50));
+                                Stock behindstock = PubMaster.Goods.GetStockBehindStockPoint(EndTrack.id, (DevStatus.CurrentPoint-50));
                                 if (behindstock != null)
                                 {
                                     TAKE_STOCK_POINT = behindstock.location;
@@ -652,7 +652,7 @@ namespace simtask
                                 }
                                 break;
                             case SimCarrierSortStepE.取货完成获取卸货位置:
-                                Stock infrontstock = PubMaster.Goods.GetStockInfrontStockPoint(EndTrack.id, (ushort)(DevStatus.CurrentPoint + 50));
+                                Stock infrontstock = PubMaster.Goods.GetStockInfrontStockPoint(EndTrack.id, (DevStatus.CurrentPoint + 50));
                                 if (infrontstock != null)
                                 {
                                     ushort safe = PubMaster.Goods.GetStackSafe(0, 0);
